requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
certifi==2019.6.16
Click==7.0
colorama==0.4.1
Django==2.2.3
django-material-admin==1.1.18
entrypoints==0.3
flake8==3.7.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.1
rope==0.14.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
wincertstore==0.2
wrapt==1.11.2
yapf==0.28.0
